resume verification/requirements.txt
matcher==0.2
numpy==2.3.5
pdf2image==1.17.0
pdfplumber==0.11.8
pytesseract==0.3.13
python-dotenv==1.2.1
python_docx==1.2.0
Requests==2.32.5
sentence_transformers==5.1.2
spacy==3.8.11
streamlit==1.51.0
yake==0.6.0
